Refactor notifications handling: integrate Notifications component into TabsLayout, replace UserLogsScreen with modal, and remove obsolete files.
